Updated code to support newer Qt versions (#237)

+++ UEFITool/qhexedit2/chunks.cpp
@@ -95,11 +95,11 @@ QByteArray Chunks::data(qint64 pos, qint64 maxSize, QByteArray *highlighted)
                     count = maxSize;
                 if (count > 0)
                 {
-                    buffer += chunk.data.mid(chunkOfs, (int)count);
+                    buffer += chunk.data.mid((int)chunkOfs, (int)count);
                     maxSize -= count;
                     pos += count;
                     if (highlighted)
-                        *highlighted += chunk.dataChanged.mid(chunkOfs, (int)count);
+                        *highlighted += chunk.dataChanged.mid((int)chunkOfs, (int)count);
                 }
             }
         }
@@ -176,7 +176,7 @@ qint64 Chunks::indexOf(const QByteArray &ba, qint64 from)
     for (qint64 pos=from; (pos < _size) && (result < 0); pos += BUFFER_SIZE)
     {
         buffer = data(pos, BUFFER_SIZE + ba.size() - 1);
-        int findPos = buffer.indexOf(ba);
+        int findPos = (int)buffer.indexOf(ba);
         if (findPos >= 0)
             result = pos + (qint64)findPos;
     }
@@ -194,7 +194,7 @@ qint64 Chunks::lastIndexOf(const QByteArray &ba, qint64 from)
         if (sPos < 0)
             sPos = 0;
         buffer = data(sPos, pos - sPos);
-        int findPos = buffer.lastIndexOf(ba);
+        int findPos = (int)buffer.lastIndexOf(ba);
         if (findPos >= 0)
             result = sPos + (qint64)findPos;
     }
@@ -214,8 +214,8 @@ bool Chunks::insert(qint64 pos, char b)
     else
         chunkIdx = getChunkIndex(pos);
     qint64 posInBa = pos - _chunks[chunkIdx].absPos;
-    _chunks[chunkIdx].data.insert(posInBa, b);
-    _chunks[chunkIdx].dataChanged.insert(posInBa, char(1));
+    _chunks[chunkIdx].data.insert((int)posInBa, b);
+    _chunks[chunkIdx].dataChanged.insert((int)posInBa, char(1));
     for (int idx=chunkIdx+1; idx < _chunks.size(); idx++)
         _chunks[idx].absPos += 1;
     _size += 1;
@@ -241,8 +241,8 @@ bool Chunks::removeAt(qint64 pos)
         return false;
     int chunkIdx = getChunkIndex(pos);
     qint64 posInBa = pos - _chunks[chunkIdx].absPos;
-    _chunks[chunkIdx].data.remove(posInBa, 1);
-    _chunks[chunkIdx].dataChanged.remove(posInBa, 1);
+    _chunks[chunkIdx].data.remove((int)posInBa, 1);
+    _chunks[chunkIdx].dataChanged.remove((int)posInBa, 1);
     for (int idx=chunkIdx+1; idx < _chunks.size(); idx++)
         _chunks[idx].absPos -= 1;
     _size -= 1;
